Move visualizeRootMargin from library to example folder. Updated docs to not mention visualizeRootMargin.
