+++ readme.md
@@ -27,60 +27,28 @@ yarn add @kaliber/use-is-intersecting
 ```jsx
 export default function App() {
   const [root, setRoot] = React.useState(null)
-  const { ref: isIntersectingRef, isIntersecting } = useIsIntersecting({ root, rootMargin })
-  const { ref: wasIntersectingRef, wasIntersecting } = useWasIntersecting({ root, rootMargin })
-
-  const rootMarginStyles = visualizeRootMargin(rootMargin)
+  const { ref: isIntersectingRef, isIntersecting } = useIsIntersecting({ root, rootMargin: '-10%' })
+  const { ref: wasIntersectingRef, wasIntersecting } = useWasIntersecting({ root, rootMargin: '-10%' })
 
   return (
     <div>
       <div className={styles.scrollWrapper}>
-        <div className={styles.rootMargin} style={rootMarginStyles} />
-
         <div ref={setRoot} className={styles.scroll}>
           <div ref={isIntersectingRef}>
-            {isIntersecting ? 'Is in intersection root element' : 'Is not in intersection root element'}
+            {isIntersecting 
+              ? 'Is in intersection root element'  
+              : 'Is not in intersection root element'}
           </div>
 
           <div ref={wasIntersectingRef} >
-            {wasIntersecting ? 'Has been in intersection root element' : 'Has not yet been in intersection root element'}
+            {wasIntersecting 
+              ? 'Has been in intersection root element' 
+              : 'Has not yet been in intersection root element'}
           </div>
         </div>
       </div>
     </div>
   )
-}
-```
-
-To use the rootMargin visualiser, make sure the component you apply these styles on wraps the root element and has the same dimensions (check the example folder).
-```css
-.rootMargin {
-  position: relative;
-  pointer-events: none;
-
-  &::before,
-  &::after {
-    content: '';
-    position: absolute;
-  }
-
-  &::before {
-    top: 0;
-    bottom: 0;
-    left: inherit;
-    right: inherit;
-    border-left: 1px solid lime;
-    border-right: 1px solid orange;
-  }
-
-  &::after {
-    left: 0;
-    right: 0;
-    top: inherit;
-    bottom: inherit;
-    border-top: 1px solid lime;
-    border-bottom: 1px solid orange;
-  }
 }
 ```
